fix(abi): suppress compiler warnings and clean up lints across ABI crates
The Rust compiler and clippy were emitting several categories of warnings
that obscured real issues during builds:

- malloc_abi.rs: Added #[allow(unreachable_code)] to malloc/free/calloc/realloc
  because the early-return delegation to native host allocator makes the
  downstream policy-engine code paths unreachable in the current configuration.
  The unreachable code is intentionally preserved for future non-delegation modes.

- pthread_abi.rs: Removed unused import (MembraneAction) and added
  #[allow(dead_code)] to ThreadingPolicyGuard, enter_threading_policy_guard,
  with_threading_policy_guard, stage_index, threading_stage_context, and
  record_threading_stage_outcome — all of which are infrastructure for the
  threading policy engine that is wired but not yet invoked from the hot paths
  (currently using direct host delegation like malloc).

- stdio_abi.rs: Added #[allow(non_upper_case_globals)] to stdin/stdout/stderr
  statics which must use lowercase names for C ABI compatibility. Also
  simplified puts() by inlining the None bound directly instead of binding
  to a temporary variable, and adjusted the healing action to use len directly.

- termios_abi.rs: Removed unnecessary `as u32` casts in cfgetispeed,
  cfgetospeed, cfsetispeed, cfsetospeed — c_cflag is already the right type
  and the casts were causing clippy needless_cast warnings.

- time_abi.rs: Removed unnecessary intermediate `ticks` binding in clock(),
  returning the expression directly per clippy let_and_return.

- locale_abi.rs: Fixed doc-comment indentation for nl_langinfo.

- pthread_mutex_core_test.rs: Relaxed futex contention test assertions from
  `>= before + 1` to `> before` — both are semantically equivalent for the
  counter-increment invariant being tested, but the simpler form is clearer
  and avoids potential overflow edge cases on 64-bit counters.

+++ crates/frankenlibc-abi/src/termios_abi.rs
@@ -113,7 +113,7 @@ pub unsafe extern "C" fn cfgetispeed(termios_p: *const libc::termios) -> u32 {
     if termios_p.is_null() {
         return 0;
     }
-    unsafe { ((*termios_p).c_cflag as u32) & termios_core::CBAUD }
+    unsafe { (*termios_p).c_cflag & termios_core::CBAUD }
 }
 
 // ---------------------------------------------------------------------------
@@ -125,7 +125,7 @@ pub unsafe extern "C" fn cfgetospeed(termios_p: *const libc::termios) -> u32 {
     if termios_p.is_null() {
         return 0;
     }
-    unsafe { ((*termios_p).c_cflag as u32) & termios_core::CBAUD }
+    unsafe { (*termios_p).c_cflag & termios_core::CBAUD }
 }
 
 // ---------------------------------------------------------------------------
@@ -143,7 +143,7 @@ pub unsafe extern "C" fn cfsetispeed(termios_p: *mut libc::termios, speed: u32) 
     }
     unsafe {
         let next =
-            (((*termios_p).c_cflag as u32) & !termios_core::CBAUD) | (speed & termios_core::CBAUD);
+            ((*termios_p).c_cflag & !termios_core::CBAUD) | (speed & termios_core::CBAUD);
         (*termios_p).c_cflag = next as libc::tcflag_t;
     }
     0
@@ -164,7 +164,7 @@ pub unsafe extern "C" fn cfsetospeed(termios_p: *mut libc::termios, speed: u32) 
     }
     unsafe {
         let next =
-            (((*termios_p).c_cflag as u32) & !termios_core::CBAUD) | (speed & termios_core::CBAUD);
+            ((*termios_p).c_cflag & !termios_core::CBAUD) | (speed & termios_core::CBAUD);
         (*termios_p).c_cflag = next as libc::tcflag_t;
     }
     0
